+++ c/index-0.0.1/src/inverter.c
@@ -8,6 +8,7 @@
 #include <limits.h>
 #include "global.h"
 #include "inverter.h"
+#include "sort.h"
 
 /* maximum size of internal index in terms */
 #define MEMORY_SIZE		(2000000)
@@ -64,7 +65,7 @@ void inverter_free(Inverter_t * v)
 /* allocate memory for inverter */
 void inverter_alloc(Inverter_t * v)
 {
-	v->size = prime(FILL_RATIO * MAX_COUNT);
+	v->size = (uint32_t)prime(FILL_RATIO * MAX_COUNT);
 	v->maxpool = 2 * MEMORY_SIZE;
 	v->count = 0;
 	v->record = (char **) malloc(v->size * sizeof(char *));
@@ -201,5 +202,5 @@ int inverter_write(Inverter_t * v, FILE * fp)
 
 int inverter_isfull(Inverter_t * v)
 {
-	return (v->count > MAX_COUNT || v->ppool - v->pool > v->maxpool);
+	return (v->count > MAX_COUNT || v->ppool-v->pool > (int)v->maxpool);
 }
